Refactor telescope components for improved structure and readability and delete useless code.
